requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Cython==3.0.10
dataclasses-json==0.5.14
diskcache==5.6.3
Distance==0.1.3
fairseq==0.12.2
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
g2p-en==2.1.0
greenlet==3.0.3
huggingface-hub==0.23.0
hydra-core==1.0.7
idna==3.7
importlib_metadata==7.1.0
inflect==7.2.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langchainplus-sdk==0.0.20
langsmith==0.1.59
llama_cpp_python==0.2.75
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
omegaconf==2.0.6
openapi-schema-pydantic==1.2.4
orjson==3.10.3
packaging==23.2
portalocker==2.8.2
protobuf==5.26.1
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
sacrebleu==2.4.2
safetensors==0.4.3
sounddevice==0.4.6
SQLAlchemy==2.0.30
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboardX==2.6.2.2
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
tqdm==4.66.4
transformers==4.40.2
typeguard==4.2.1
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.2
